app/mqttrw/requirements.txt
Jinja2==3.0.1
paho-mqtt==1.5.1
itsdangerous==2.0.1
Werkzeug==2.0.1
redis
